# faster {IntMap,IntSet} size functions