@treeowl 2017-03-29
@wrengr, these numbers look convincing, if surprising. Do you agree? Also, it looks like `foldl'` for `IntSet` may need some work.