@mikeplus64 2017-03-29
A couple little tests ^

I use `List.length . toList` as well as the foldl' thing because the original IntSet prop_size used it instead of foldl'.