@treeowl 2017-03-29
Oops. Sorry about that.

On Wed, Mar 29, 2017 at 2:49 AM, Mike Ledger <notifications@github.com>
wrote:

> *@mikeplus64* commented on this pull request.
> ------------------------------
>
> In tests/intset-properties.hs
> <https://github.com/haskell/containers/pull/421#discussion_r108601766>:
>
> > @@ -261,8 +261,10 @@ prop_isSubsetOf a b = isSubsetOf a b == Set.isSubsetOf (toSet a) (toSet b)
>  prop_isSubsetOf2 :: IntSet -> IntSet -> Bool
>  prop_isSubsetOf2 a b = isSubsetOf a (union a b)
>
> -prop_size :: IntSet -> Bool
> -prop_size s = size s == List.length (toList s)
> +prop_size :: IntSet -> Property
> +prop_size s = sz === foldl' (\i _ -> i + 1) (0 :: Int) s .&&.
> +              sz === List.length (toList s)
> +  where sz = size s
>
>
> Thankfully, it was already there. I was a little (happily) confused for a
> moment as well.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/haskell/containers/pull/421#discussion_r108601766>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABzi_dcEPHYqxQQiCDsk85Ghs91TSbGDks5rqf7dgaJpZM4Msf5e>
> .
>