@treeowl 2017-03-30
Never mind about `foldl'` for `IntSet`. That's obviously not a good way to implement `size`. I wasn't paying enough attention.